+++ test/test_encoding.rb
@@ -8,15 +8,13 @@ describe 'Encoding' do
     should 'support i18n content' do
       memcached do |dc|
         key = 'foo'
-        bad_key = utf8 = 'ƒ©åÍÎ'
+        utf_key = utf8 = 'ƒ©åÍÎ'
 
         assert dc.set(key, utf8)
         assert_equal utf8, dc.get(key)
 
-        # keys must be ASCII
-        assert_raises ArgumentError, /illegal character/ do
-          dc.set(bad_key, utf8)
-        end
+        dc.set(utf_key, utf8)
+        assert_equal utf8, dc.get(utf_key)
       end
     end
 
@@ -27,15 +25,6 @@ describe 'Encoding' do
         assert_equal 'bar', dc.get(key)
         sleep 2
         assert_equal nil, dc.get(key)
-      end
-    end
-
-    should 'not allow non-ASCII keys' do
-      memcached do |dc|
-        key = 'fooƒ'
-        assert_raises ArgumentError do
-          dc.set(key, 'bar')
-        end
       end
     end